[CMake] Ensure link_directories is invoked early enough.
The documentation at
https://cmake.org/cmake/help/latest/command/link_directories.html
explicitely states :

The command will apply only to targets created after it is called, so
fix the invocation order in the CMakeLists.txt.

While being there, remove the 'PRIVATE' keyword which does not apply to
this flavour of the linker command setup.

This does fix building tarmac-gui-browser on Mac and remove spurious
warnings about non-existing xyz/PRIVATE directories.