[Docs] Change blockquotes to description lists.
In ReST, an indented paragraph separated by a blank line from a prior
unindented one renders to HTML as a <blockquote>. If you don't put the
blank line in between, you get a <dl> in which the unindented
paragraph appears as a <dt> and the indented one as a <dd>, which is
generally more appropriate for what I intended to communicate.

The difference between the two wasn't at all obvious in my offline
test renderings of the manual, but Github's online renderer puts
styling on <blockquote> that makes it very clear that I got the wrong
markup elements. I think this patch fixes them all.