Reverse the order of 'bytes' in RegisterEvent.
We receive register values in a big-endian form from trace files, and
ultimately need to convert them into little-endian to store in our
internal representation in the index. That conversion was previously
happening in index.cpp after it receives a RegisterEvent. Now it's
happening in parser.cpp before it _sends_ the same event. In other
words, the 'bytes' field in RegisterEvent itself is now taken to be
little-endian (matching the internal representation) instead of
big-endian.

This will make the semantics simpler when we need to specify partial
register updates.