Remove the optional 'value' field in RegisterEvent.
The only receiver of a RegisterEvent that actually used it (apart from
in parsertest diagnostics) was the one that tracks updates to sp in
the indexer. There's already a perfectly good system in the indexer
for reading the current value of a register back out of the current
memtree, so it's at least as easy to just use that, so that there's
only one kind of RegisterEvent and it always behaves the same.