Add an offset field to RegisterEvent. (NFC)
This permits the data type to represent a partial update to a
register, overwriting only some of its bytes, and not necessarily
starting at the LSB. The receiving code in the indexer uses the
provided offset to adjust the location of its memtree update, and also
takes into consideration the possibility that ev.bytes may not equal
the full size of the register.

Nothing actually uses this facility yet. It will in the next commit,
but this commit has a lot of boring updates to the test output file,
so I separate it out so that the interesting part can stand alone.