Remove size parameter from lookup_reg_name.
It existed in order to disambiguate register names that sometimes
appear in Tarmac traces meaning different things (the key example is
that "SP" can mean AArch32 r13, or AArch64 xsp). But it was awkward
for it to be a mandatory part of the API, because sometimes you don't
know in advance what size you expect the register to be. For example,
Browser::TraceView::lookup_register (in the context of user-provided
arithmetic expressions) had to loop over all possible sizes when the
user referred to a register.

Now lookup_reg_name _never_ takes a size, so Browser::TraceView can do
something much simpler. The previous size-dependent special cases in
lookup_reg_name are now part of the Tarmac parser code, since that's
the only place those ambiguities are expected to show up.

The payoff for this is that we can also support a form of register
update seen in the output of some RTL simulators, in which a 128-bit
vector register is emitted as multiple 8-digit hex numbers separated
by spaces. We now know in advance how big we expect any register to be
(except SP, which is handled as a special case), and can keep parsing
data from the input trace line until we have enough.

One existing parser regression test which has changed its result is
the one involving FPCR being listed as a 64-bit register update. Now
the update returns just the low 32 bits of the data, which is all that
will fit into our representation of the register anyway.