Support partial register updates.
Some RTL simulators, such as the one for the Cortex-A53, can emit
register update lines in which only _part_ of a register is changed,
and the unchanged bytes are shown as "--" in place of hex digits. The
case of this that I've seen involves a write to a double-precision FP
register, which the trace represents as a partial update of the vector
register it overlaps.