Fix searching for the next visit to the current PC.
We were beginning the search from the start rather than the end of the
current logical node. Since pressing the 'next' button in the browser
treats the PC in question as the one _after_ the current node, this
caused us to just re-find the current position every time.

This bug must have been introduced when I switched over the internal
primary key for a trace node from its timestamp to its line number
(after discovering that the former wasn't always unique), because the
timestamp doesn't have a range of valid values. Also, in passing,
fixed an outdated comment from before that change.