Change strategy for calling cmake 'install' command.
Previously, the 'browser' and 'tools' subdirectories each appended the
names of some targets to the list variable 'installed_tools', and then
the top-level CMakeLists.txt handed the completed list to the CMake
install command.

It turns out that this fails on Ubuntu 18.04's system cmake (3.10.2),
which objects for some reason to the install targets not being defined
in the same cmake subdirectory script that called 'install' itself. I
didn't notice, because our CI uses the standard Github 18.04 runner,
which provides a much more up-to-date cmake in place of the system
one.

I've disposed of the list variable, and replaced it with a macro
wrapping the 'install' command (and providing the same alternative
definitions as before depending on the cmake version). Now the build
works on vanila 18.04.

Fixes #4.