Add a wxWidgets version of tarmac-gui-browser.
This is intended to take over completely from the GTK version. Its
main advantage is that it can build on Windows much more easily: GTK3
on Windows is hard enough to build that I haven't managed it at all
yet! So this is an actually practical approach to making the GUI
browser able to run on Windows.

For the moment, this is in an experimental state, which is why it has
the name tarmac-wx-browser, and the GTK one still has the name
tarmac-gui-browser. Once the wxWidgets version has had some more
testing, I hope to switch it over to pride of place, and then retire
the GTK one completely.