Initial support for Tarmac generated by Cortex-M4 RTL.
This flavour of Tarmac is approximately in the Fast Models style
(IT/IS instruction records, contiguous rather than diagrammatic memory
transfers), but with some awkward differences that aren't clearly
signalled: the CPU mode field is missing from instruction records, and
the data in memory transfers is given in memory order rather than
logical order.

As far as I can tell, these changes manage to support this format
without either breaking the existing test suite or requiring an
up-front mode setting for the Tarmac parser. But it's precarious: the
clues from which I deduce which syntax and semantics to expect don't
really have much to do with the syntax and semantics _differences_, so
it's all a bit more 'recognise the fingerprint of a particular
implementation' than I'd like.

However, it does work, and this change also adds some Cortex-M4 lines
to the parser test suite with their expected output. So if we need to
completely rethink things later, we'll be able to make sure the new
approach still works.