[wx] Leave position at top of an unfolded function.
When a TraceWindow's current position is on a folded-up function call,
the physical node (relative to which all the registers and memory are
displayed) is the last instruction of the folded-up region, i.e. the
function return. But the _visible_ node is the much earlier function
_call_. The effect is to make it seem, while the call is folded up, as
if that one call instruction enacted all the effects of the whole
subroutine.

When the function call is unfolded using unfold_ui_action(), the wrong
one of those nodes took precedence: the physical node remained the
same, and the visible node was updated to match it. As a result, the
cursor ends up at the _end_ of the unfolded function. I think a better
default behaviour is to have it end up at the start, so that you can
start stepping through the function.