[wx] Left and Right arrow keys now fold and unfold.
The detailed behaviour is inspired by GUI tree views, such as the one
in regedit. In an unfolded function, pressing Left will move the
current position to the start of the function (just after the call
instruction), unless that was where the current position was _already_
(e.g. after a previous press of Left), in which case the function will
be folded up. Meanwhile, with the cursor just after a folded function,
pressing Right will unfold it (and, thanks to the previous commit,
leave you at the top of the unfolded section).