[wx] Fix assertion failure in context menu setup.
When right-clicking on a folded function, an assertion failure could
come up with a message

  ../src/common/menucmn.cpp(310): assert "wxIsStockID(GetId())" failed
  in SetItemLabel(): A non-stock menu item with an empty label?

Confusing, since it didn't _look_ as if I was creating any menu item
with an empty label! It turns out this was because I was making the
'section heading' for a folded function (really a disabled wxMenuItem)
by passing a std::string containing a nontrivial UTF-8 character, and
in that situation, the default conversion to wxString seems to
silently generate the empty string. Fixed by adding an explicit
wxString::FromUTF8, similarly to the way I was already doing it for
unfolded functions.

(Also, while I'm at it, expanded the UTF-8 en dash in my string
literal into explicit \x sequences, so that it's much more obvious
that it's a dangerous character.)