Add a new 'Reporter' abstraction.
This replaces all the previous pieces of code that would output
diagnostics and progress reports straight to standard error: fatal OS
errors, trace file parsing errors, reports on indexing progress and on
whether (and why) indexing was being performed at all.

Reporter itself is an abstract base class, and there's a global
unique_ptr<Reporter> variable which allows all parts of the code to
access the current application's single instance of it.

NFC in this commit: the only implementation of Reporter as yet is the
CommandLineReporter defined in lib/misc.cpp, which implements the
entire API by doing exactly the same things that I just replaced with
calls to that API.