Remove the GTK version of tarmac-gui-browser.
It was always my intention to switch over to this version once it was
ready: the wxWidgets version is much easier to compile for Windows
(I've tried, but never managed to compile any GTK3 program
successfully on Windows at all), and the native C++ nature of
wxWidgets suits this code base a lot better than the ad-hoc C++ I
layered on top of GTK. Not only that, but wxWidgets can back-end on to
GTK3 in any case, so even if someone especially wanted the GTK3 widget
set, this wouldn't stop them from having it.

I've been using the wxWidgets version a bit now, and I think it's now
at feature parity with the GTK version, and doesn't have any obvious
bugs. So it's time to switch over, and then further bug fixes and
features can be added in one GUI browser only.

The binary 'tarmac-wx-browser' is now renamed to tarmac-gui-browser,
and the GTK version that previously had that name is no more.

That also means there's no real need for the tiny source file
main-browser.cpp any more, because that was shared between the curses
and GTK browsers but not the wxWidgets one. So that main() now lives
in curses.cpp ony.