[wx] Use wxSizerFlags objects.
I found this idiom in the wxWidgets example code, and it seems to be a
more legible approach to saying what you want the controls in your
sizer to be doing.

This rework fixes the very large space between the wxStaticText and
the wxTextCtrl in the dialog that prompts for a memory address to
display.