[wx] Make MemPromptDialog's Cancel button actually cancel!
I completely forgot to pay attention to the 'bool ok' parameter in
mem_prompt_dialog_ended. As a result, the Cancel button was being
treated just like OK, which meant that not only would a memory window
be opened whether you liked it or not, but you couldn't even get rid
of the dialog box until you'd entered a valid memory address in it!