[wx] Provide a startup dialog box.
Previously, tarmac-gui-browser expected to be run from the command
line in the same kind of way as all the other CLI tools, with at least
a trace file name and optionally other settings. So if you tried to
run it with an empty command line, e.g. by double-clicking its
executable in a file manager or launching it from a Start menu or via
a GUI shortcut, it would simply complain that you hadn't given it a
valid command line, and terminate.

Now, in that situation, it will put up a preliminary dialog box in
which you can select a trace file to browse, and optionally fill in a
couple of other controls corresponding to command-line options. So
it's now feasible to do GUI-style launching of this GUI application.

The startup box is only presented if the command line _is_ empty.
Otherwise, the same command-line parsing is done as before. So any
command line you could previously use without an error will still work.