Warning fix: struct/class mismatches.
In a couple of places, a type was predeclared as 'class Foo;', but
when it came to the implementation, I found it more convenient to
write 'struct' instead of 'class'. According to clang, "this is valid,
but may result in linker errors under the Microsoft C++ ABI".

It's no difficulty to fix: from my point of view writing 'struct'
instead of 'class' is just a way to avoid having to write 'public:'
straight after the open brace. In one case, I've done that. In the
other case it seemed more appropriate to make a friendship relation
between the class in question (TarmacLineParserImpl) and
TarmacLineParser which it implements, since the point is that that's
the only class that ought to be able to see the contents.