Warning fixes that only happen at NDEBUG.
When asserts are compiled out, some variables become unused (in that I
only assigned into them so I could check them with an assertion
afterwards), and one code path starts giving a warning about not
returning a value from the containing function (because it had an
assert indicating that that path should never be reached anyway).

Squashed the former warnings via '(void)variable;', and the latter
with a phony return value that won't be used anyway.