CMakeLists: support cross-compiling for Windows.
Compiling for Windows on Linux almost works out of the box, with an
appropriate cmake toolchain file, except that tarmac-gui-browser
doesn't get built because the call to find_package(wxWidgets) gets
confused at the unexpected combination of Windows-ness and Unix-ness.

It can be worked around by manually resetting a few variables, though.
This patch made it work for me, and shouldn't activate unwantedly in
any other circumstances.