Warning fix: use the Windows secure CRT functions.
The MSVC headers warn about the standard C functions fopen, localtime
and asctime being 'deprecated', and suggests using the _s variants
instead. Normally my instinct is to squash those warnings via
_CRT_SECURE_NO_WARNINGS, but in this case, there are so few of them
that we can actually fix them. So I've added a set of wrapper
functions in misc.hh whose Windows implementation calls the 'secure'
functions, and the Unix implementation still uses the standard
versions.

(I don't think it makes much difference, in fact. As I understand it
the difference between standard fopen and fopen_s is the default
permissions on files they create - but I only open files for reading
here. asctime and localtime are not thread safe, but this program is
single-threaded.)