Warning fix: reinterpret_cast that changes size.
In wx.cpp I needed to store an 'unsigned' in a private-data field of
type 'void *', and did it via a reinterpret_cast. That provokes a
warning from clang, on the basis that it's confusing for
reinterpret_cast to change the size of a thing.

Easily enough fixed by first static_casting the input integer to one
the same size as the pointer, and then reinterpret_casting *that* to
an actual pointer. That explicitly expresses the semantics I wanted,
and doesn't give a warning.