expr.cpp: external support for constant expressions.
This exposes a function to constructing an ExprPtr containing a
constant expression evaluating to a constant you already have in
machine-readable form, without having to annoyingly format it into a
parseable string and re-parse it.

Also, I've exposed a test method you can use to find out whether an
expression _is_ constant.