New menu option to open a MemoryWindow on the stack.
Of all the variable base addresses you might possibly want for your
MemoryWindow, I think this is the single most important one: the base
address of the window is the stack pointer, so that as you step in and
out of function calls, you can always see the current stack frame in
your memory view.

It's so important that it seems worth its own menu option, so you
don't have to keep manually entering the same address expression.