MemoryWindow: store an ExprPtr internally. (NFCI)
Alongside the existing field 'start_addr', we now store an optional
ExprPtr, representing a parsed expression that can be evaluated to
recompute the window's start address. If this expression exists, then
we re-evaluate it every time the window's memroot changes, i.e.
whenever the linked TraceWindow moves to a new position in the trace.

The idea is that it will be possible to specify a variable base
address for memory windows, as an arithmetic expression based on
things like register values which can vary with the current trace
position.

Alongside start_addr_expr, there's also start_addr_exprstr, which is a
string representation of the same expression. This is used to

This commit is groundwork only: the new expr and exprstr fields are
added, and if they exist then the expression is re-evaluated at the
right time and used as the initial contents of the edit box in the
'change address' dialog, but nothing actually sets them up yet. So,
for the moment, no functional change: all current MemoryWindows still
work the old way.