MemoryWindow: allow new windows with non-constant expressions.
This is the other half of the change in the previous commit: you can
now enter a non-constant expression in the Ctrl-M dialog box before a
MemoryWindow is created at all.

To do that, I changed the constructor of MemoryWindow to take a small
class I made up that either holds a constant, or a parsed expression
and its string form.