MemoryWindow: allow change to non-constant address expression.
When you already have a MemoryWindow in existence, you can now enter a
register-dependent expression in the address box in the toolbar. After
you've done that, the window's base address will automatically reset
to the new value of that expression whenever you change the trace file
position that the window is looking at (e.g. move the current position
in the linked TraceWindow, if there is one).

In between those events, you can still scroll the window up and down
to look around the current position, but the scrolled position is
transient - it will reset to the exact value of the address expression
at the next change of trace file position.

This change allows you to enter a nontrivial address expression into
an _existing_ MemoryWindow, but not to create a MemoryWindow from
scratch with anything but a constant expression. That involves more
code upheaval, and will come in the next commit.