lib/CMakeLists: add ARCHIVE DESTINATION as well as LIBRARY.
If you build libtarmac as a shared library, the LIBRARY DESTINATION
clause in the install command tells cmake where to install it. But if
you build it as a static library, you need ARCHIVE DESTINATION instead.

The type of the libtarmac library is left unspecified in CMakeLists,
which means it can be controlled by -DBUILD_SHARED_LIBS=(ON|OFF) on
the cmake command line. So the install command should work either way.