[cmake] Add a mode to force building all applications.
Our CMake files allows to not build applications by default when some of
their dependencies are missing. This means some applications may never
be built and/or tested with our CI if there is a missing dependency.
This patch provides a stricter mode where the building of all
applications is forced by ensuring cmake will fail at the configure
stage if there is a dependency missing.

The strict mode can be invoked from the command line with:

  > cmake -DFORCE_BUILDING_ALL_APPS:BOOL=ON ...

If FORCE_BUILDING_ALL_APPS is not set on the command line, then it will
default to OFF, reverting to the user friendly behaviour.