Support partial reg updates via "R regname<hibit:lobit>".
Fixes #9, which reported this syntax as coming from a Neoverse-E1
Cycle Model.

This is an alternative syntax for specifying an update of only part of
a register, alongside the already-supported approach of showing some
of the bytes of the register contents as "--" to indicate that they're
unchanged from their previous value.

In principle, this syntax could specify a slice of a register right
down to the bit level, since the interval bounds are specified in
bits. But TTU's internal data structure for a register update event
only has byte granularity, so currently we reject any update in this
format that can't be represented as a whole number of bytes. (The only
case so far observed in the wild is the one shown in parsertest.txt,
specifying the upper or lower 64 bits of a vector register.)