Add a RegisterId representation of pc.
pc is not actually stored in the register address space (hence it's
declared as having size 0, so it doesn't even _consume_ any of the
address space). So it didn't previously have a RegisterId, because for
most purpses in this code, it's not really the same kind of thing as
other registers.

But in the context of expression evaluation, it _is_ interchangeable
with other registers, in that they all live in the same namespace. And
I'm about to reorganise the evaluator, and it will be useful to be
able to refer to anything looking like a register using the same data
type.