Make a Parser class wrapping the expression parser.
The recursive-descent parse functions are now all methods of 'Parser',
rather than being free functions.

NFC: the purpose is so that the Parser class can store further state
in future, but for the moment, all it stores is the Lexer & that was
previously passed as an argument to all the parse functions.