Refactor symbol and register lookup in expression parsing.
Previously, the parse stage of expression evaluation would preserve
every symbolic name in the form of a ScopedId or UnscopedId AST node,
containing the symbol name still in the form of a string. At actual
evaluation time, that would be looked up in the register name system
and/or the ELF symbol table, which might fail if the name didn't exist.

As a result, there was no way to tell the difference between the
failure modes 'this name doesn't exist at all' and 'it exists, but
refers to a register whose value isn't known at this point in the
trace'.

I've now reorganised it so that the names are looked up at parse time,
so that 'name doesn't even exist' errors can be reported much earlier,
at the same time as literal _parse_ errors like mismatched brackets.
The parse phase now terminates with scope disambiguation already
finished; register names translated from strings into RegisterId form;
and symbols from the ELF image already translated into their actual
numeric values (since those don't change as you move around the
trace). So now an evaluation-time failure to look something up can
_only_ be because the value of a register is not available at a given
trace position.

In particular, this means that if you press Ctrl-M in the GUI browser
to open a new memory window, and enter an expression into the dialog
box that contains a nonexistent name, the error can be reported
immediately and you get a chance to correct the expression.