Reorder lines in exprtest.txt.
In the previous commit, I added some extra tests at the bottom of the
file. That wasn't really where they belonged, but it made the diff
easier to read. But now that's done, I'm reorganising the file to put
related tests together again, so _this_ separate commit contains the
unavoidable huge churn in exprtest.ref when all the line numbers
change.