tarmac-browser: support for auto-moving memory windows.
This adds the same feature in the curses Tarmac browser that commits
e40bfec5346e6ac and 5640ed1640c1494 implemented in the GUI: now if you
enter an expression at the 'm' prompt that includes time-variable
values (i.e. register contents), the browser will remember the
expression it evaluated, and re-evaluate it every time the locked
trace window changes position, so that you can have a memory window
follow (for example) the stack pointer.

When a memory window is following a movable expression in this way,
the window's status line will show it (paralleling the edit box in the
toolbar of the GUI memory window).

This makes it a little awkward to set up a _static_ memory window
pointing at the _current_ value of a register. It's already a little
awkward in the GUI browser too. I'll fix both in the next commit.