Implement an 'unfollow' action in both browsers.
Now, if you have a memory window following a variable address
expression, it's easy to make it stop. In the Curses browser, there's
a new keypress 'F' (rationale: for 'follow', with the shift key
indicating negation). In the GUI browser, you can simply delete the
entire contents of the address expression edit box, and this will be
taken to mean 'behave as if I had re-entered the address I'm currently
looking at'.