[wx] Fix provenance queries in RegisterWindow.
Right-clicking a register in the register window and selecting 'Go to
last write to this register' was going to the wrong place.
Specifically, it was misidentifying _which_ register to find the last
write to, because I'd got confused about which fields of LogicalPos
are used for what in that particular window type. In multi-column
register layouts (i.e. most of them), if you clicked on a register
that wasn't in the leftmost column, you'd get a provenance query for
the leftmost register in that row instead.

(This bug was introduced in the port to wxWidgets, because I changed
the semantics of LogicalPos as part of that process. The previous
version of this code was correct in the GTK browser.)