[wx] Propagate logical location update when unfolding.
When you have the cursor placed just after a folded-up function call,
and unfold it (e.g. by pressing Right or clicking the + sign), then
the logical position of the cursor changes from 'just after the entire
call' to 'just after the call instruction enters the called function'.
This ought to cause all the linked register and memory subviews to
update. But it didn't, because I missed out an update_location call.