Add missing rebase in update_memtree_from_read.
Whenever we allocate more space from the memory-mapped index file,
it's possible that the file itself might have to be extended, and then
re-mmapped, potentially at a different address. So any pointers into
the mapped area that the calling code is holding should be 'rebased'
by recomputing them from the new base offset of the mapped region.

The main loop in update_memtree_from_read() had two potential
allocation events (allocating a block of space to put the newly read
data in, and the tree insert operation that allocates a tree node
structure), and it was only rebasing its 'subroot' pointer after one
of them. So if the other happened to trigger a re-mmap, there'd be
trouble.