Make index file generation crash-safe.
Suppose the TTU indexer terminates in an uncontrolled way during the
indexing process. (For example, a crash such as the one fixed in the
previous commit, or the user hitting ^C, or the disk filling up, or
any other cause.)

Then something _looking_ like an index file will be left on disk, with
the correct magic number, and a timestamp newer than the trace file.
So the next time you run a TTU tool on that trace, it will decide the
index file doesn't need to be regenerated - but it's not actually a
valid index file, and will cause a messy explosion!

To fix this, I've added a new FLAG_COMPLETE to the flags word in the
file header, which is only set at the very end of index generation.
And the initial check of the file's magic number also checks for
FLAG_COMPLETE, so that it will detect an incomplete file and know it
can't be safely used.

This is a file format change, so I've also bumped the version field in
the magic number. As a result, any older files which don't set this
flag will be regenerated anyway, because of the version update.