GUI browser: support double- and triple-click selection.
Like terminal windows, a double-click selects a 'word' or some other
kind of semantic unit around the click location, and a triple-click
selects a whole line. And after you start a selection in one of those
ways, dragging the selection bounds further will continue to select a
whole number of words or lines.

But in this application, we can do a bit better than the terminal's
interpretation of 'word' as 'longest contiguous sequence of some
subset of the character set'. Trace display windows have the syntax
highlighting from the parse to refer to, and register windows have the
display classes written by format_reg. So we can regard a word as a
contiguous sequence of characters with the same highlighting class,
i.e. _semantically_ connected in a more specific sense. In particular,
double-clicking the instruction disassembly selects the whole
instruction.

One particularly awkward part of this was finding out how to
distinguish single-, double- and triple-clicks in wxWidgets. I ended
up having to redo most of the work myself, because wxWidgets has a
special event type for 'double-click' but not for 'triple'. However,
the reimplementation honours the local OS's configurable settings
where possible.