Separate character display type for register names.
Previously, in a register display, the register name and the following
'=' sign were treated as the same class of text. Now they're different
from each other.

In the Curses front end, this makes no difference, because the two
classes are mapped to the same terminal colour. In the wxWidgets GUI
I've made them slightly different shades of dark green, because we
have more colours conveniently available.

But the real payoff isn't the visual difference: it will come in a
couple of commits' time, when those character classes are used to
determine selection bounds.