Parser: new highlighting class for whitespace between tokens.
Previously, whitespace was highlighted as HL_NONE, unless it was in
the middle of the instruction's disassembly, in which case it received
the same HL_DISASSEMBLY as the rest of the instruction - including
trailing whitespace at the end of the line.

Now I've made up a new highlighting class HL_SPACE to apply to most
whitespace. It covers any whitespace not returned by the lexer as a
token at all, and it also covers trailing whitespace after the
disassembly. But whitespace _within_ the disassembly I've left as it
is: that's still HL_DISASSEMBLY.

There should be no visual difference in this. But the next commit will
use it for selection bounds.