Fix a few compiler warnings.
In browse.cpp, one of the if statements matching a class of branch
instruction used an expression of the form 'a && b || c && d', where
the ones around it had extra parens '(a && b) || (c && d)'. The former
got a warning from clang (benign, since it was interpreted the same as
the latter anyway).

Also added a cast-to-void of a variable that only exists to be used in
an assertion (so it becomes unused under -DNDEBUG conditions).