[IndexReader] Constification.
The methods for reading from the tarmac file can actually be const as
they do modify the state of IndexReader from a user perspective. It's
just an implementation detail that there is file stream opened in which
the cursor is moving. The position after the read is definitely a state,
but it's irrelevant (and internal) to the IndexReader.