[wx] Fix assertion on selecting an address in MemoryWindow.
If you select one of the memory addresses in the leftmost column,
which don't represent actual bytes of memory, then an assertion fails
in addr_from_valid_logpos, whose caller get_region_under_pos didn't
check that before trying to find out what address you'd clicked on.

The caller of that in turn was update_mouseover, which just wanted to
know what memory - if any - the mouse was either hovering over or
selecting. So 'none' is a perfectly sensible answer in this case.