Fix two inappropriate uses of 'mod_time'.
SeqOrderPayload nodes can have the same mod_time, if the Tarmac trace
has sufficiently low time resolution that multiple instructions are
shown with identical timestamps.

So it's always a mistake to test the mod_time field to decide whether
two SeqOrderPayloads are the same one. Testing trace_file_firstline is
the right thing, because that is _never_ the same between distinct
nodes.

A couple of sites in browse.cpp contained tests of mod_time, from
before I figured that out. In particular this could sometimes block
unfolding of a folded-up section of trace, if it caused
physline_range_for_folded_function_after() to spuriously return false
to TraceWindow::FunctionRange::set_to_callee().