Change the typedef 'Time' to be 64-bit.
Up-to-date versions of Arm Fast Models have switched to writing Tarmac
timestamps in units of picoseconds instead of clock cycles, which
means the numbers are much bigger, and can easily overflow the 32-bit
time field in SeqOrderPayload.

(In fact, that was what caused the unfolding failure I mention in the
previous commit message: in the Tarmac file where I experienced the
problem, the timestamps of the two nodes in question _were_ different
- but they were both bigger than 2^32, and had been clipped to the
same value during parsing.)

This is an index-file format change, so I've bumped the magic number
version.

Also fixed a type declaration in calltree.cpp, in which
generate_flame_graph_recurse() expected a map<string, Time>, but its
caller generate_flame_graph was passing a map<string, unsigned>. Those
used to be the same, but now are not. Easily fixed.