[Argparse] Teach Argparse to support via-files.
Some applications can be invoked with a number of arguments exceeding
the shell or OS capabilities. The common solution to this problem is to
pass the arguments in a file, the so-called via-file.

This commit modifies the implementation of Argparse to support this use
case. The private arguments vector is now a deque and the user code can
modify arguments, prepending or appending words to arguments, while
parsing the command line.