Ignore mapping symbols in ELF images.
These symbols are there a static disassembler which parts of the file
they should be disassembling as Arm, Thumb or data (in AArch32), or as
code or data (in AArch64). They're extremely common and often have
identical names, so they're never useful as a means of identifying a
particular address - even the numeric address would be better!

So this patch ignores them when reading the ELF file, and then we'll
never unhelpfully decode an address as $a.14+0x1234 or similar nonsense.