Add some stress tests of the btod module.
I'm reminded of it by Arnaud's reorganisation last night, and now I
look at it, I see that its list of test cases are all the really
obvious things like min and max float values, or 1 plus or minus an
ULP, or π.

A technique has been well known for decades to find the FP values
closest to a rounding boundary of the output decimal representation,
so as to verify that the implementation correctly computes enough
digits internally to round even those the right way. btod.cpp computes
in exact big integers, so it ought to pass all those tests, but none
of them were included in the test suite.

This patch adds a selection of those values: the hardest rounding
cases at both the minimum and maximum exponents (testing dealing with
the largest intermediate integer values), and the hardest _overall_
rounding cases (that is, over all possible exponents, the ones that
differ from the rounding boundary by the least amount in ULP). Also,
included some cases that _exactly_ hit the rounding boundary, testing
the round-to-even tie breaking rule.