Turn MMapFile into a small class hierarchy.
The abstract base class is 'Arena', which can be any kind of thing
that keeps a large contiguous block of address space and is able to
allocate things from the end of it, resizing it as needed.

MMapFile is now just one subclass of Arena. There's another one,
MemArena, which isn't backed by a disk file at all and just maintains
a contiguous region of ordinary memory through the much more obvious
medium of malloc and realloc.

This should make it possible in future for tools to not bother keeping
their index on disk, if they don't really need to. Currently, no
functional change.

In particular, the previous MMapFile::alloc() method has been broken
into two pieces. The alloc() method itself lives in the base class,
but delegates to the concrete subclass the task of resizing the
arena (because that's the part where you have to change the size of a
disk file, or call realloc, or whatever).