Index class: store a whole TracePair, not separate filenames.
It's a historical oddity that the index file name is given to the
Index at construction time, but the trace file name is passed to the
parse_tarmac_file method. Since both those filenames live in the
TracePair structure, it's easier to just pass the whole thing at
construction time.