Make --show-progress-meter do something again.
It was apparently broken by commit 4b84b68bdd4099e (introducing the
Reporter abstraction), because that caused the progress-reporting code
in index.cpp to stop conditioning on its show_progress_meter
parameter, and instead just pass progress reports unconditionally to
the Reporter.

That makes sense, because in a GUI context the progress meter is
always shown, so index.cpp really does need to do that. But ever since
then, the command-line Reporter has been unconditionally _not_
printing a progress meter, because I forgot to put in the compensating
line of code in TarmacUtilityBase which uses the show_progress_meter
member variable to set the verbosity of the Reporter.