Make TarmacUtilityBase::updateIndexIfNeeded not static.
It's always called from a (sub)class instance, and its parameters
include four fields copied from the calling instance, so there's no
real point in having it be static. Instead, it's now const, so that it
can't accidentally overwrite those parameters.

(In particular, it does modify its own copy of the 'indexing' field,
and it's important that it shouldn't accidentally modify the original,
because in a multi-file tool indexing==Auto will have to be handled
independently per trace.)