Add option to store the trace index in memory.
This modifies all the existing tools to add a --memory-index option,
which in turn sets a flag in TracePair to indicate that instead of an
index file name it's holding a MemArena that will contain the index.

At the moment I expect that this is only really useful as a test of
all the refactoring I've just done. I can imagine situations in
general where saving temporary index files on disk would be
inconvenient because you don't have a writable filesystem available,
but it's hard to imagine someone in that kind of emergency _also_
having a desperate need to browse Tarmac files :-)

My longer-term aim is that tools that do one-off analysis should be
able to override the default setting of the index_on_disk flag so that
they _never_ store an index on disk. But that won't be useful until
there are also options to cut down the amount of stuff stored in the
index, so that it only takes up the space really needed by each
particular tool.